% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_company_names.R
\name{add_company_names}
\alias{add_company_names}
\title{Find the company behind the operator ID}
\usage{
add_company_names(vector, dataset = oildata::pipelines_2010)
}
\arguments{
\item{vector}{Vector with operator IDs to be replaced.}

\item{dataset}{Dataset with operator names.}
}
\value{
Vector, where the operator IDs with matches are replaced with the name of the company.
}
\description{
Find the company behind the operator ID
}
\examples{
add_company_names(c(300, 26049, 25146))
}
