% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_utils.R
\name{fix_commodities}
\alias{fix_commodities}
\title{Recode a commodity vector to be standardized between the different datasets.}
\usage{
fix_commodities(x)
}
\arguments{
\item{x}{Vector with commodities to be standardized.}
}
\value{
Cleaned up vector as factor.
}
\description{
Recode a commodity vector to be standardized between the different datasets.
}
\examples{
oildata:::fix_commodities(incidents_2002$commodity)
}
