% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topic_loadings.R
\name{get_topic_loadings}
\alias{get_topic_loadings}
\title{Get topic loadings aggregated to the organization}
\usage{
get_topic_loadings(df = oildata::incidents, variable = volume)
}
\arguments{
\item{df}{Dataset with topic loadings.}

\item{variable}{Variable used for weighting the loadings.}
}
\value{
The dataset with identity columns and loadings only.
}
\description{
Get topic loadings aggregated to the organization
}
\examples{
get_topic_loadings()
}
