% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_groups.R
\name{handle_ma}
\alias{handle_ma}
\title{Handle M&As in a dataframe}
\usage{
handle_ma(df, summary_parsing, m_as, by_cols = vars(ID, year, commodity))
}
\arguments{
\item{df}{Dataframe for which to aggregate companies to organizational level.}

\item{summary_parsing}{A list of new column names and quoted functions (with quo) to be passed into summarize.}

\item{m_as}{The M&As section of the m_as dataset in this package.}

\item{by_cols}{Columns by which to aggregate.}
}
\description{
Aggregates observations where an M&A has taken place. Is called by oildata::consolidate_groups,
after company groups are handled.
}
