% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_utils.R
\docType{data}
\name{incident_consolidation}
\alias{incident_consolidation}
\title{Rules for adding together different incidents when aggregating by organization and/or year etc.}
\format{
A list with the rowname, and the summary operation.
\describe{
  \item{incident_consolidation}{List that holds entry for each column as separate object.}
}
}
\source{
Wirtten for the purpose of handling operations in this package. Source of the processed data:
  \url{https://www.phmsa.dot.gov/data-and-statistics/pipeline/distribution-transmission-gathering-lng-and-liquid-accident-and-incident-data}
}
\usage{
incident_consolidation
}
\description{
Rules for adding together different incidents when aggregating by organization and/or year etc.
}
\examples{
\dontrun{
 incident_consolidation
}
}
\keyword{datasets}
