% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidents.R
\docType{data}
\name{incidents}
\alias{incidents}
\title{Merged dataset of all incidents.}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{year}{}
  \item{ID}{}
  \item{commodity}{}
  \item{significant}{}
  \item{serious}{}
  \item{installation_year}{Year item installed that involved in accident.}
  \item{cause}{}
  ...
}}
\usage{
incidents
}
\description{
Merged dataset of all incidents.
}
\keyword{datasets}
