% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidents_2002.R
\docType{data}
\name{incidents_2002}
\alias{incidents_2002}
\title{Annual data on oil spills in the US from 2004 to 2010}
\format{A data frame with columns:
\describe{
 \item{DATAFILE_AS_OF}{}
 \item{significant}{Identify if record meets the significant criteria or not: If there was fatality, injury, fire, explosion, total property damage $50K or more in 1984 dollars, non-HVL loss >= 50bbls, HVL loss >= 5bbls, then SIGNIFICANT=’YES’, else SIGNIFICANT=’NO’. See also: \url{https://www.phmsa.dot.gov/sites/phmsa.dot.gov/files/docs/pdmpublic_incident_page_allrpt.pdf}}
 \item{serious}{Identify if record meets the SERIOUS criteria or not: If there was fatality or injury then SERIOUS = ‘YES’ else SERIOUS = ’NO’.}
 \item{report_type}{}
 \item{year}{Year of the incident report.}
 \item{ID}{Unique operator ID.}
 \item{RPTID}{}
 \item{DOR}{}
 \item{OWNER_OPERATOR_ID}{}
 \item{name}{Name of the pipelines operator at the time of the incident report.}
 \item{OPSTREET}{}
 \item{OPCITY}{}
 \item{OPCOUNTY}{}
 \item{OPSTATE}{}
 \item{OPZIP}{}
 \item{IHOUR}{}
 \item{IDATE}{}
 \item{LATITUDE}{}
 \item{LONGITUDE}{}
 \item{ACCITY}{}
 \item{ACCOUNTY}{}
 \item{ACSTATE}{}
 \item{ACZIP}{}
 \item{MPVST}{}
 \item{SURVNO}{}
 \item{TELRN}{}
 \item{TELDT}{}
 \item{PPPRP}{}
 \item{EMRPRP}{}
 \item{ENVPRP}{}
 \item{OPCPRP}{}
 \item{OPCPRPO}{}
 \item{PRODPRP}{}
 \item{OOPRP}{}
 \item{OOPPRP}{}
 \item{OOPPRPO}{}
 \item{cost}{Total costs ($).}
 \item{SPILLED}{}
 \item{COMM}{}
 \item{commodity}{The type of commodity being spilled (Crude, HVL, etc.).}
 \item{SPUNIT_TEXT}{}
 \item{LOSS}{}
 \item{RECOV}{}
 \item{cause}{Apparent Cause of the Incident (self-reported).}
 \item{LINE_SEG}{}
 \item{IFED}{}
 \item{INTER}{}
 \item{on_offshore}{Whether the oil spill occured onshore or offshore.}
 \item{OFFAREA}{}
 \item{BNUMB}{}
 \item{OFFST}{}
 \item{OCS}{}
 \item{OPPROP}{}
 \item{PIPEROW}{}
 \item{HCA}{}
 \item{HCADESC}{}
 \item{SYSPRT_TEXT}{}
 \item{SYSPRTO}{}
 \item{FAIL_OC_TEXT}{}
 \item{FAIL_OCO}{}
 \item{installation_year}{Year item installed that involved in accident.}
 \item{INC_PRS}{}
 \item{MOP}{}
 \item{OPRS}{}
 \item{LRTYPE_TEXT}{}
 \item{LRTYPEO}{}
 \item{LEAK_TEXT}{}
 \item{PUNC_DIAM}{}
 \item{RUPTURE_TEXT}{}
 \item{RUPLN}{}
 \item{PROPLN}{}
 \item{UBLKVM}{}
 \item{UBLKVA}{}
 \item{UBLKVR}{}
 \item{UBLKVC}{}
 \item{DBLKVM}{}
 \item{DBLKVA}{}
 \item{DBLKVR}{}
 \item{DBLKVC}{}
 \item{SEGISO}{}
 \item{VLVDIST}{}
 \item{SEGCONF}{}
 \item{INLINE_TEXT}{}
 \item{DRHRMF}{}
 \item{DRHRMFY}{}
 \item{DRLRMF}{}
 \item{DRLRMFY}{}
 \item{DRUT}{}
 \item{DRUTY}{}
 \item{DRGEO}{}
 \item{DRGEOY}{}
 \item{DRCAL}{}
 \item{DRCALY}{}
 \item{DRCRK}{}
 \item{DRCRKY}{}
 \item{DRHARD}{}
 \item{DRHARDY}{}
 \item{DROTH}{}
 \item{DROTHY}{}
 \item{NPS}{}
 \item{WALLTHK}{}
 \item{SPEC}{}
 \item{SMYS}{}
 \item{SEAM}{}
 \item{VALVE}{}
 \item{MANU}{}
 \item{MANYR}{}
 \item{LOCLK_TEXT}{}
 \item{LOCLKO}{}
 \item{DEPTH_COV}{}
 \item{FATAL}{}
 \item{EFAT}{}
 \item{NFAT}{}
 \item{GPFAT }{}
 \item{INJURE}{}
 \item{EINJ}{}
 \item{NINJ}{}
 \item{GPINJ}{}
 \item{SHUTDOWN}{}
 \item{SHUTDAY}{}
 \item{SHUTHR }{}
 \item{SHUTMIN}{}
 \item{IGNITE}{}
 \item{EXPLO}{}
 \item{EVAC}{}
 \item{EVACNO}{}
 \item{EVAC_REASON_TEXT }{}
 \item{STHH}{}
 \item{STMN}{}
 \item{FISH}{}
 \item{BIRDS}{}
 \item{TERRESTRIAL}{}
 \item{SOIL}{}
 \item{SOIL_YRD }{}
 \item{IMPACT}{}
 \item{REMEDIAL}{}
 \item{RSURFACE}{}
 \item{RGROUND}{}
 \item{RSOIL}{}
 \item{RVEG}{}
 \item{RWILD }{}
 \item{WATER}{}
 \item{AMT_IN_WATER}{}
 \item{OCEAN}{}
 \item{SURFACE}{}
 \item{GROUNDW}{}
 \item{DRINK}{}
 \item{DRINKSRC_TEXT}{}
 \item{COMP_BASED}{}
 \item{DETECTED_TEXT}{}
 \item{DETECTEDO}{}
 \item{DURLEAK_DAY}{}
 \item{DURLEAK_HR}{}
 \item{CAUSE_DETAILS}{}
 \item{PIPE_COAT_TEXT}{}
 \item{VIS_EXAM_TEXT}{}
 \item{VIS_EXAMO}{}
 \item{COR_CAUSE_TEXT}{}
 \item{COR_CAUSEO}{}
 \item{PROT}{}
 \item{CPYR}{}
 \item{PREV_DAM}{}
 \item{PREV_DAM_YR }{}
 \item{PREV_DAM_MO}{}
 \item{PREV_DAM_UK}{}
 \item{EARTH_MOVE_TEXT}{}
 \item{EARTH_MOVEO}{}
 \item{FLOODS_TEXT }{}
 \item{FLOODSO}{}
 \item{TEMPER_TEXT}{}
 \item{TEMPERO}{}
 \item{THIRD_PARTY_GRP_TEXT}{}
 \item{THIRD_PARTY_TYPE_TEXT}{}
 \item{THIRD_PARTY_TYPEO}{}
 \item{EXCAV_TYPE_TEXT}{}
 \item{EXCAV_ON}{}
 \item{EXCAV_LAST_CONTACT}{}
 \item{NOTIF }{}
 \item{NOTIF_DATE}{}
 \item{NOTIF_RCVD_TEXT}{}
 \item{MARKED}{}
 \item{TEMP_MARK_TEXT}{}
 \item{PERM_MARK}{}
 \item{ACC_MARK_TEXT}{}
 \item{MKD_IN_TIME}{}
 \item{FIRE_EXPLO_TEXT}{}
 \item{PIPE_BODY_TEXT}{}
 \item{PIPE_BODYO }{}
 \item{COMPONENT_TEXT}{}
 \item{COMPONENTO}{}
 \item{JOINT_TEXT}{}
 \item{JOINTO}{}
 \item{BUTT_TEXT}{}
 \item{BUTTO}{}
 \item{FILLET_TEXT}{}
 \item{FILLETO}{}
 \item{PIPE_SEAM_TEXT }{}
 \item{PIPE_SEAMO}{}
 \item{FAIL_TYPE_TEXT}{}
 \item{CONS_DEF_TEXT}{}
 \item{PIPE_DAMAGE}{}
 \item{PRS_TEST }{}
 \item{TEST_YR}{}
 \item{TEST_MO}{}
 \item{TEST_DAY}{}
 \item{TEST_MED_TEXT}{}
 \item{TEST_MEDO}{}
 \item{TEST_TP}{}
 \item{TEST_PRS}{}
 \item{MALFUNC_TEXT}{}
 \item{MALFUNCO}{}
 \item{THREADS_TEXT}{}
 \item{THREADSO}{}
 \item{SEAL_TEXT }{}
 \item{SEALO}{}
 \item{IO_TYPE_TEXT}{}
 \item{IO_TYPEO}{}
 \item{IO_DRUG}{}
 \item{IO_ALCO}{}
 \item{MISC }{}
 \item{UNKNOWN_TEXT}{}
 \item{PNAME}{}
 \item{PHONE}{}
 \item{PEMAIL}{}
 \item{narrative}{Written description of the incident.}
 \item{PPPRPCURRENT}{}
 \item{EMRPRPCURRENT}{}
 \item{ENVPRPCURRENT}{}
 \item{OPCPRPCURRENT}{}
 \item{PRODPRPCURRENT}{}
 \item{OOPRPCURRENT}{}
 \item{OOPPRPCURRENT}{}
 \item{TOTAL_COST_IN84}{}
 \item{TOTAL_COST_CURRENT}{}
 \item{MAP_CAUSE}{}
 \item{MAP_SUBCAUSE}{}
 \item{SPILL_TYPE_CATEGORY}{}
 \item{volume}{Volume of oil spilled (in gallons)}
}}
\source{
United States Department of Transportation (DOT) Pipeline and
  Hazardous Materials Safety Administration (PHMSA).
  \url{https://www.phmsa.dot.gov/data-and-statistics/pipeline/distribution-transmission-gathering-lng-and-liquid-accident-and-incident-data}
   (right side bar).
}
\usage{
incidents_2002
}
\description{
Distribution, Transmission & Gathering, LNG, and Liquid Accident and Incident Data.
}
\details{
Title 49 of the Code of Federal Regulations (49 CFR Parts 191, 195) requires
pipeline operators to submit incident reports within 30 days of a pipeline
incident or accident. The CFR defines accidents and incidents, as well as
criteria for submitting reports to the Office of Pipeline Safety. The
following kinds of information are collected:
* Key report information
* In-depth location information
* Facility information
* Operating information
* Drug and Alcohol information
* Cause of the accident/incident
Specific information includes the time and location of the incident(s),
number of any injuries and/ or fatalities, commodity spilled/gas released,
causes of failure and evacuation procedures. The reports are used for
identifying long- and short-term trends at the national, state and
operator-specific levels. The frequency, causes, and consequences of the
incidents provide insight into the safety metrics currently used by PHMSA,
state partners, and other pipeline safety stakeholders, including the
pipeline industry and general public. PHMSA also uses the data for inspection
planning and risk assessment.
}
\examples{
\dontrun{
 incidents_2002
}
}
\keyword{datasets}
