% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidents_2010.R
\docType{data}
\name{incidents_2010}
\alias{incidents_2010}
\title{Annual data on oil spills in the US from 2010 onward}
\format{A data frame with columns:
\describe{
 \item{DATAFILE_AS_OF}{}
 \item{significant}{Identify if record meets the significant criteria or not: If there was fatality, injury, fire, explosion, total property damage $50K or more in 1984 dollars, non-HVL loss >= 50bbls, HVL loss >= 5bbls, then SIGNIFICANT=’YES’, else SIGNIFICANT=’NO’. See also: \url{https://www.phmsa.dot.gov/sites/phmsa.dot.gov/files/docs/pdmpublic_incident_page_allrpt.pdf}}
 \item{serious}{Identify if record meets the SERIOUS criteria or not: If there was fatality or injury then SERIOUS = ‘YES’ else SERIOUS = ’NO’.}
 \item{ipe}{Impacting People or the Environment (IPE) – when commodity (A8) is crude oil, refined petroleum products, of biofuel, if either criterion 1 or 2 below is met, the accident counts as IPE :  1.  Regardless of Location of Accident (B10): Fatality (A12) greater than zero; or Injury requiring in-patient hospitalization (A13) greater than zero; or Ignition (A15) = Yes; or Explosion (A16) = Yes; or Evacuation (A17) greater than zero; or Wildlife impact (D1) = Yes; or Water contamination (D5a) = Ocean/Seawater, Groundwater, or Drinking water; or Public/Non-Operator Private Property Damage (D8a) greater than zero 2.For Location of Accident (B10) not “TOTALLY CONTAINED ON OPERATOR CONTROLLED PROPERTY”: Unintentional Release Volume (A9) greater than or equal to 5 gallons AND HCA (D7) = Yes; or Unintentional Release Volume (A9) greater than or equal to 5 barrels AND HCA (D7) = No; or Water contamination (D5a) = Surface; or Soil contamination (D2) = Yes }
 \item{integrity_assessment_target}{Integrity Assessment Target – accidents Impacting People or the Environment (IPE) and one of these causes: Corrosion, Pipeline/Weld Material Failure, Failure of Previously Damage Pipe – caused by Excavation Damage or Other Outside Force Damage.}
 \item{operations_maintenance_target}{Operation & Maintenance Target - accidents Impacting People or the Environment (IPE) and one of these causes: Equipment Failure, Incorrect Operation, 1st & 2nd Party Excavation Damage, 3rd Party Excavation Damage with Root Cause = Locating Practices Not Sufficient.}
 \item{REPORT_RECEIVED_DATE}{}
 \item{year}{Year of the incident report.}
 \item{REPORT_NUMBER}{}
 \item{SUPPLEMENTAL_NUMBER}{}
 \item{REPORT_TYPE}{}
 \item{ID}{Unique operator ID.}
 \item{name}{Name of the pipelines operator at the time of the report year.}
 \item{OPERATOR_STREET_ADDRESS}{}
 \item{OPERATOR_CITY_NAME}{}
 \item{OPERATOR_STATE_ABBREVIATION}{}
 \item{OPERATOR_POSTAL_CODE}{}
 \item{LOCAL_DATETIME}{}
 \item{LOCATION_LATITUDE}{}
 \item{LOCATION_LONGITUDE}{}
 \item{NRC_RPT_NUM}{}
 \item{NRC_RPT_DATETIME}{}
 \item{commodity}{The type of commodity being spilled (Crude, HVL, etc.).}
 \item{COMMODITY_SUBTYPE}{}
 \item{COMMODITY_DETAILS}{}
 \item{BLEND_DETAILS}{}
 \item{BIO_DIESEL_DETAILS}{}
 \item{UNINTENTIONAL_RELEASE_BBLS}{}
 \item{INTENTIONAL_RELEASE_BBLS}{}
 \item{RECOVERED_BBLS}{}
 \item{FATALITY_IND}{}
 \item{NUM_EMP_FATALITIES}{}
 \item{NUM_CONTR_FATALITIES}{}
 \item{NUM_ER_FATALITIES}{}
 \item{NUM_WORKER_FATALITIES}{}
 \item{NUM_GP_FATALITIES}{}
 \item{FATAL}{}
 \item{INJURY_IND}{}
 \item{NUM_EMP_INJURIES}{}
 \item{NUM_CONTR_INJURIES}{}
 \item{NUM_ER_INJURIES}{}
 \item{NUM_WORKER_INJURIES}{}
 \item{NUM_GP_INJURIES}{}
 \item{INJURE}{}
 \item{SHUTDOWN_DUE_ACCIDENT_IND}{}
 \item{SHUTDOWN_EXPLAIN}{}
 \item{SHUTDOWN_DATETIME}{}
 \item{RESTART_DATETIME}{}
 \item{STILL_SHUTDOWN_IND}{}
 \item{IGNITE_IND}{}
 \item{EXPLODE_IND}{}
 \item{NUM_PUB_EVACUATED}{}
 \item{INCIDENT_IDENTIFIED_DATETIME}{}
 \item{ON_SITE_DATETIME}{}
 \item{ON_OFF_SHORE}{}
 \item{ONSHORE_STATE_ABBREVIATION}{}
 \item{ONSHORE_POSTAL_CODE}{}
 \item{ONSHORE_CITY_NAME}{}
 \item{ONSHORE_COUNTY_NAME}{}
 \item{DESIGNATED_LOCATION}{}
 \item{DESIGNATED_NAME}{}
 \item{PIPE_FAC_NAME}{}
 \item{SEGMENT_NAME}{}
 \item{FEDERAL}{}
 \item{LOCATION_TYPE}{}
 \item{INCIDENT_AREA_TYPE}{}
 \item{INCIDENT_AREA_SUBTYPE}{}
 \item{INCIDENT_AREA_DETAILS}{}
 \item{DEPTH_OF_COVER}{}
 \item{CROSSING}{}
 \item{BRIDGE_CROSSING_IND}{}
 \item{BRIDGE_TYPE}{}
 \item{RAILROAD_CROSSING_IND}{}
 \item{RAILROAD_TYPE}{}
 \item{ROAD_CROSSING_IND}{}
 \item{ROAD_TYPE}{}
 \item{WATER_CROSSING_IND}{}
 \item{WATER_TYPE}{}
 \item{WATER_NAME}{}
 \item{WATER_DEPTH}{}
 \item{WATER_SUBTYPE}{}
 \item{OFF_WATER_DEPTH}{}
 \item{OFF_ACCIDENT_ORIGIN}{}
 \item{OFFSHORE_STATE_ABBREVIATION}{}
 \item{OFF_INSTATE_AREA}{}
 \item{OFF_INSTATE_BLOCK}{}
 \item{OFFSHORE_COUNTY_NAME}{}
 \item{OFF_OCS_AREA}{}
 \item{OFF_OCS_BLOCK}{}
 \item{OFF_AREA_ACCIDENT_TYPE}{}
 \item{PIPE_FACILITY_TYPE}{}
 \item{SYSTEM_PART_INVOLVED}{}
 \item{SYSTEM_SUBPART_INVOLVED}{}
 \item{ITEM_INVOLVED}{}
 \item{PIPE_TYPE}{}
 \item{PIPE_DIAMETER}{}
 \item{PIPE_WALL_THICKNESS}{}
 \item{PIPE_SMYS}{}
 \item{PIPE_SPECIFICATION}{}
 \item{PIPE_SEAM_TYPE}{}
 \item{PIPE_SEAM_DETAILS}{}
 \item{PIPE_MANUFACTURER}{}
 \item{PIPE_MANUFACTURE_YEAR}{}
 \item{PIPE_COATING_TYPE}{}
 \item{PIPE_COATING_DETAILS}{}
 \item{WELD_SUBTYPE}{}
 \item{WELD_DETAILS}{}
 \item{VALVE_TYPE}{}
 \item{VALVE_MAINLINE_TYPE}{}
 \item{VALVE_MAINLINE_DETAILS}{}
 \item{VALVE_MANUFACTURER}{}
 \item{VALVE_MANUFACTURE_YEAR}{}
 \item{TANK_VESSEL_SUBTYPE}{}
 \item{ITEM_INVOLVED_DETAILS}{}
 \item{TANK_VESSEL_DETAILS}{}
 \item{installation_year}{Year item installed that involved in accident.}
 \item{MATERIAL_INVOLVED}{}
 \item{MATERIAL_DETAILS}{}
 \item{RELEASE_TYPE}{}
 \item{PUNCTURE_AXIAL}{}
 \item{PUNCTURE_CIRCUM}{}
 \item{LEAK_TYPE}{}
 \item{LEAK_TYPE_OTHER}{}
 \item{RUPTURE_ORIENT}{}
 \item{RUPTURE_DETAILS}{}
 \item{RUPTURE_WIDTH}{}
 \item{RUPTURE_LENGTH}{}
 \item{RELEASE_TYPE_DETAILS}{}
 \item{WILDLIFE_IMPACT_IND}{}
 \item{FISH_AQUATIC_IMPACT_IND}{}
 \item{BIRDS_IMPACT_IND}{}
 \item{TERRESTRIAL_IMPACT_IND}{}
 \item{SOIL_CONTAMINATION}{}
 \item{LONG_TERM_ASSESSMENT}{}
 \item{REMEDIATION_IND}{}
 \item{SURFACE_WATER_REMED_IND}{}
 \item{GROUNDWATER_REMED_IND}{}
 \item{SOIL_REMED_IND}{}
 \item{VEGETATION_REMED_IND}{}
 \item{WILDLIFE_REMED_IND}{}
 \item{WATER_CONTAM_IND}{}
 \item{OCEAN_SEAWATER_IND}{}
 \item{SURFACE_CONTAM_IND}{}
 \item{GROUNDWATER_CONTAM_IND}{}
 \item{DRINKING_WATER_CONTAM_IND}{}
 \item{PRIVATE_WELL_CONTAM_IND}{}
 \item{PUBLIC_WATER_CONTAM_IND}{}
 \item{AMOUNT_RELEASED}{}
 \item{REL_WATER_NAME}{}
 \item{COULD_BE_HCA}{}
 \item{COMMODITY_REACHED_HCA}{}
 \item{COMMERCIALLY_NAV_IND}{}
 \item{COMMERCIALLY_NAV_YES_NO}{}
 \item{HIGH_POP_IND}{}
 \item{HIGH_POP_YES_NO}{}
 \item{OTHER_POP_IND}{}
 \item{OTHER_POP_YES_NO}{}
 \item{USA_DRINKING_IND}{}
 \item{USA_DRINKING_YES_NO}{}
 \item{USA_ECOLOGICAL_IND}{}
 \item{USA_ECOLOGICAL_YES_NO}{}
 \item{EST_COST_OPER_PAID}{}
 \item{EST_COST_GAS_RELEASED}{}
 \item{EST_COST_PROP_DAMAGE}{}
 \item{EST_COST_EMERGENCY}{}
 \item{EST_COST_ENVIRONMENTAL}{}
 \item{EST_COST_OTHER}{}
 \item{EST_COST_OTHER_DETAILS}{}
 \item{total_cost}{Total estimated property damage (sum of above)}
 \item{ACCIDENT_PSIG}{}
 \item{MOP_PSIG}{}
 \item{ACCIDENT_PRESSURE}{}
 \item{PRESSURE_RESTRICTION_IND}{}
 \item{EXCEED_RESTRICTION_IND}{}
 \item{PHMSA_RESTRICTION_IND}{}
 \item{PART_C_QUESTION_2_IND}{}
 \item{UPSTREAM_VALVE_TYPE_IND}{}
 \item{DOWNSTREAM_VALVE_TYPE_IND}{}
 \item{LENGTH_SEGMENT_ISOLATED}{}
 \item{INTERNAL_INSPECTION_IND}{}
 \item{DIAMETER_CHANGE_IND}{}
 \item{UNSUITABLE_MAINLINE_IND}{}
 \item{TIGHT_MITERED_IND}{}
 \item{OTHER_RESTRICTIONS_IND}{}
 \item{EXTRA_THICK_WALL_IND}{}
 \item{OTHER_INSPECTION_IND}{}
 \item{INTERNAL_INSPECTION_DETAILS}{}
 \item{OPERATION_COMPLICATIONS_IND}{}
 \item{EXCESSIVE_DEBRIS_IND}{}
 \item{LOW_OP_PRESSURE_IND}{}
 \item{LOW_FLOW_IND}{}
 \item{INCOMPAT_COMMOD_IND}{}
 \item{OTHER_COMPLICATIONS_IND}{}
 \item{INSPECT_COMP_DETAILS}{}
 \item{PIPELINE_FUNCTION}{}
 \item{SCADA_IN_PLACE_IND}{}
 \item{SCADA_OPERATING_IND}{}
 \item{SCADA_FUNCTIONAL_IND}{}
 \item{SCADA_DETECTION_IND}{}
 \item{SCADA_CONF_IND}{}
 \item{CPM_IN_PLACE_IND}{}
 \item{CPM_OPERATING_IND}{}
 \item{CPM_FUNCTIONAL_IND}{}
 \item{CPM_DETECTION_IND}{}
 \item{CPM_CONF_IND}{}
 \item{ACCIDENT_IDENTIFIER}{}
 \item{ACCIDENT_DETAILS}{}
 \item{OPERATOR_TYPE}{}
 \item{INVESTIGATION_STATUS}{}
 \item{INVESTIGATION_STATUS_DETAILS}{}
 \item{INVEST_SCHEDULE_IND}{}
 \item{INVEST_NO_SCHEDULE_IND}{}
 \item{INVEST_NO_SCHEDULE_IND_DETAILS}{}
 \item{INVEST_NO_CONTROL_ROOM_IND}{}
 \item{INVEST_NO_CONTROLLER_IND}{}
 \item{INVEST_INCORRECT_ACTION_IND}{}
 \item{INVEST_FATIGUE_IND}{}
 \item{INVEST_INCORRECT_PROCEDURE_IND}{}
 \item{INVEST_INCORRECT_CONTROL_IND}{}
 \item{INVEST_MAINT_IND}{}
 \item{INVEST_OTHER_IND}{}
 \item{INVEST_OTHER_IND_DETAILS}{}
 \item{EMPLOYEE_DRUG_TEST_IND}{}
 \item{CONTRACTOR_DRUG_TEST_IND}{}
 \item{NUM_EMPLOYEES_TESTED}{}
 \item{NUM_EMPLOYEES_FAILED}{}
 \item{NUM_CONTRACTORS_TESTED}{}
 \item{NUM_CONTRACTORS_FAILED}{}
 \item{cause}{Apparent Cause of the Incident (self-reported).}
 \item{CAUSE_DETAILS}{}
 \item{INTERNAL_EXTERNAL}{}
 \item{VISUAL_EXAM_RESULTS}{}
 \item{VISUAL_EXAM_DETAILS}{}
 \item{GALVANIC_CORROSION_IND}{}
 \item{ATMOSPHERE_CORROSION_IND}{}
 \item{STRAY_CURRENT_CORROSION_IND}{}
 \item{MICROBIOLOGICAL_CORROSION_IND}{}
 \item{SELECTIVE_SEAM_CORROSION_IND}{}
 \item{OTHER_CORROSION_IND}{}
 \item{CORROSION_TYPE_DETAILS}{}
 \item{FIELD_EXAM_BASIS_IND}{}
 \item{METALLURGICAL_BASIS_IND}{}
 \item{OTHER_BASIS_IND}{}
 \item{CORROSION_BASIS_DETAILS}{}
 \item{UNDERGROUND_LOCATION}{}
 \item{UNDER_CATHODIC_PROTECTION_IND}{}
 \item{CATHODIC_PRO_START_YEAR}{}
 \item{SHIELDING_EVIDENT}{}
 \item{CATHODIC_SURVEY_TYPE}{}
 \item{CP_ANNUAL_SURVEY_IND}{}
 \item{CLOSE_INTERVAL_SURVEY_IND}{}
 \item{OTHER_CP_SURVEY_IND}{}
 \item{CP_ANNUAL_SURVEY_YEAR}{}
 \item{CLOSE_INTERVAL_SURVEY_YEAR}{}
 \item{OTHER_CP_SURVEY_YEAR}{}
 \item{EXTERNALLY_COATED}{}
 \item{PRIOR_DAMAGE}{}
 \item{INT_VISUAL_EXAM_RESULTS}{}
 \item{INT_VISUAL_EXAM_DETAILS}{}
 \item{INT_CORROSIVE_COMMODITY_IND}{}
 \item{INT_WATER_ACID_IND}{}
 \item{INT_MICROBIOLOGICAL_IND}{}
 \item{INT_EROSION_IND}{}
 \item{INT_OTHER_CORROSION_IND}{}
 \item{INT_CORROSION_TYPE_DETAILS}{}
 \item{INT_FIELD_EXAM_BASIS_IND}{}
 \item{INT_METALLURGICAL_BASIS_IND}{}
 \item{INT_OTHER_BASIS_IND}{}
 \item{INT_CORROSION_BASIS_DETAILS}{}
 \item{INT_LOW_POINT_PIPE_LOC_IND}{}
 \item{INT_ELBOW_LOC_IND}{}
 \item{INT_OTHER_LOC_IND}{}
 \item{CORROSION_LOCATION_DETAILS}{}
 \item{CORROSION_INHIBITORS}{}
 \item{CORROSION_LINING}{}
 \item{CLEANING_DEWATERING}{}
 \item{CORROSION_COUPONS}{}
 \item{API_STD_OUT_OF_SERVICE_IND}{}
 \item{API_STD_OUT_OF_SERVICE_YEAR}{}
 \item{API_STD_IN_SERVICE_IND}{}
 \item{API_STD_IN_SERVICE_YEAR}{}
 \item{COR_INSP_TOOL_COLL_IND}{}
 \item{COR_MAGNETIC_FLUX_LEAKAGE_IND}{}
 \item{COR_MAG_FLUX_LEAKAGE_YEAR}{}
 \item{COR_ULTRASONIC_IND}{}
 \item{COR_ULTRASONIC_YEAR}{}
 \item{COR_GEOMETRY_IND}{}
 \item{COR_GEOMETRY_YEAR}{}
 \item{COR_CALIPER_IND}{}
 \item{COR_CALIPER_YEAR}{}
 \item{COR_CRACK_IND}{}
 \item{COR_CRACK_YEAR}{}
 \item{COR_HARDSPOT_IND}{}
 \item{COR_HARDSPOT_YEAR}{}
 \item{COR_COMBINATION_TOOL_IND}{}
 \item{COR_COMBINATION_TOOL_YEAR}{}
 \item{COR_TRANSVERSE_FIELD_IND}{}
 \item{COR_TRANSVERSE_FIELD_YEAR}{}
 \item{COR_INSPECTION_OTHER_IND}{}
 \item{COR_INSPECTION_OTHER_YEAR}{}
 \item{COR_INSPECTION_OTHER_DETAILS}{}
 \item{COR_HYDROTEST_CONDUCTED_IND}{}
 \item{COR_HYDROTEST_CONDUCTED_YEAR}{}
 \item{COR_HYDROTEST_PRESSURE}{}
 \item{DIRECT_INSPECTION_TYPE}{}
 \item{DIRECT_YES_DIG_YEAR}{}
 \item{DIRECT_YES_NO_DIG_YEAR}{}
 \item{COR_NON_DESTRUCTIVE_IND}{}
 \item{COR_RADIOGRAPHY_IND}{}
 \item{COR_RADIOGRAPHY_YEAR}{}
 \item{COR_GUIDED_WAVE_IND}{}
 \item{COR_GUIDED_WAVE_YEAR}{}
 \item{COR_HANDHELD_ULTRA_IND}{}
 \item{COR_HANDHELD_ULTRA_YEAR}{}
 \item{COR_WET_MAGNETIC_IND}{}
 \item{COR_WET_MAGNETIC_YEAR}{}
 \item{COR_DRY_MAGNETIC_IND}{}
 \item{COR_DRY_MAGNETIC_YEAR}{}
 \item{COR_NON_DEST_OTHER_IND}{}
 \item{COR_NON_DEST_OTHER_YEAR}{}
 \item{COR_NON_DEST_DETAILS}{}
 \item{NATURAL_FORCE_TYPE}{}
 \item{EARTH_SUBTYPE}{}
 \item{HEAVY_RAINS_SUBTYPE}{}
 \item{LIGHTNING_SUBTYPE}{}
 \item{TEMPERATURE_SUBTYPE}{}
 \item{NF_OTHER_DETAILS}{}
 \item{NF_EXTREME_WEATHER_IND}{}
 \item{NF_HURRICANE_IND}{}
 \item{NF_TROPICAL_STORM_IND}{}
 \item{NF_TORNADO_IND}{}
 \item{NF_OTHER_IND}{}
 \item{NF_EXTREME_WEATHER_DETAILS}{}
 \item{excavation_damage_type}{Type of damage caused by excavation party.}
 \item{EX_INSPECT_TOOL_COLLECTED_IND}{}
 \item{EX_MAGNETIC_FLUX_LEAKAGE_IND}{}
 \item{EX_MAGNETIC_FLUX_LEAKAGE_YEAR}{}
 \item{EX_ULTRASONIC_IND}{}
 \item{EX_ULTRASONIC_YEAR}{}
 \item{EX_GEOMETRY_IND}{}
 \item{EX_GEOMETRY_YEAR}{}
 \item{EX_CALIPER_IND}{}
 \item{EX_CALIPER_YEAR}{}
 \item{EX_CRACK_IND}{}
 \item{EX_CRACK_YEAR}{}
 \item{EX_HARDSPOT_IND}{}
 \item{EX_HARDSPOT_YEAR}{}
 \item{EX_COMBINATION_TOOL_IND}{}
 \item{EX_COMBINATION_TOOL_YEAR}{}
 \item{EX_TRANSVERSE_FIELD_IND}{}
 \item{EX_TRANSVERSE_FIELD_YEAR}{}
 \item{EX_INSPECTION_OTHER_IND}{}
 \item{EX_INSPECTION_OTHER_YEAR}{}
 \item{EX_INSPECTION_OTHER_DETAILS}{}
 \item{EX_BEFORE_DAMAGE}{}
 \item{EX_HYDROTEST_CONDUCTED_IND}{}
 \item{EX_HYDROTEST_CONDUCTED_YEAR}{}
 \item{EX_HYDROTEST_PRESSURE}{}
 \item{EX_DIRECT_INSPECTION_TYPE}{}
 \item{EX_DIRECT_YES_DIG_YEAR}{}
 \item{EX_DIRECT_YES_NO_DIG_YEAR}{}
 \item{EX_NON_DESTRUCTIVE_IND}{}
 \item{EX_RADIOGRAPHY_IND}{}
 \item{EX_RADIOGRAPHY_YEAR}{}
 \item{EX_GUIDED_WAVE_IND}{}
 \item{EX_GUIDED_WAVE_YEAR}{}
 \item{EX_HANDHELD_ULTRA_IND}{}
 \item{EX_HANDHELD_ULTRA_YEAR}{}
 \item{EX_WET_MAGNETIC_IND}{}
 \item{EX_WET_MAGNETIC_YEAR}{}
 \item{EX_DRY_MAGNETIC_IND}{}
 \item{EX_DRY_MAGNETIC_YEAR}{}
 \item{EX_NON_DEST_OTHER_IND}{}
 \item{EX_NON_DEST_OTHER_YEAR}{}
 \item{EX_NON_DEST_OTHER_DETAILS}{}
 \item{PRIOR_NOTIFICATION_IND}{}
 \item{ONE_CALL_SYSTEM_IND}{}
 \item{EXCAVATOR_IND}{}
 \item{CONTRACTOR_IND}{}
 \item{LANDOWNER_IND}{}
 \item{NOTIFY_CGA_DIRT}{}
 \item{PUBLIC_ROW_IND}{}
 \item{PUBLIC_SUBTYPE}{}
 \item{PRIVATE_ROW_IND}{}
 \item{PRIVATE_SUBTYPE}{}
 \item{PIPELINE_EASEMENT_ROW_IND}{}
 \item{POWER_TRANSMISSION_ROW_IND}{}
 \item{RAILROAD_ROW_IND}{}
 \item{PUBLIC_UTIL_EASEMENT_ROW_IND}{}
 \item{FEDERAL_LAND_ROW_IND}{}
 \item{DATA_NOT_COLLECTED_ROW_IND}{}
 \item{UNKNOWN_ROW_IND}{}
 \item{EXCAVATOR_TYPE}{}
 \item{EXCAVATOR_EQUIPMENT}{}
 \item{WORK_PERFORMED}{}
 \item{ONE_CALL_NOTIFIED_IND}{}
 \item{ONE_CALL_TICKET_NUM}{}
 \item{ONE_CALL_CENTER_NAME}{}
 \item{LOCATOR_TYPE}{}
 \item{VISIBLE_MARKS}{}
 \item{FACILITIES_MARKED}{}
 \item{SERVICE_INTERRUPTION}{}
 \item{SERVICE_INTERRUPTION_HOURS}{}
 \item{ROOT_CAUSE}{}
 \item{ONE_CALL_SUBTYPE}{}
 \item{LOCATING_SUBTYPE}{}
 \item{EXCAVATION_SUBTYPE}{}
 \item{ROOT_CAUSE_OTHER}{}
 \item{OUTSIDE_FORCE_TYPE}{}
 \item{VEHICLE_SUBTYPE}{}
 \item{OSF_HURRICANE_IND}{}
 \item{OSF_TROPICAL_STORM_IND}{}
 \item{OSF_TORNADO_IND}{}
 \item{OSF_HEAVY_RAINS_IND}{}
 \item{OSF_OTHER_WEATHER_IND}{}
 \item{OSF_OTHER_WEATHER_DETAILS}{}
 \item{OSF_INSPECT_TOOL_COLLECTED_IND}{}
 \item{OSF_MAGNETIC_FLUX_LEAKAGE_IND}{}
 \item{OSF_MAGNETIC_FLUX_LEAKAGE_YEAR}{}
 \item{OSF_ULTRASONIC_IND}{}
 \item{OSF_ULTRASONIC_YEAR}{}
 \item{OSF_GEOMETRY_IND}{}
 \item{OSF_GEOMETRY_YEAR}{}
 \item{OSF_CALIPER_IND}{}
 \item{OSF_CALIPER_YEAR}{}
 \item{OSF_CRACK_IND}{}
 \item{OSF_CRACK_YEAR}{}
 \item{OSF_HARDSPOT_IND}{}
 \item{OSF_HARDSPOT_YEAR}{}
 \item{OSF_COMBINATION_TOOL_IND}{}
 \item{OSF_COMBINATION_TOOL_YEAR}{}
 \item{OSF_TRANSVERSE_FIELD_IND}{}
 \item{OSF_TRANSVERSE_FIELD_YEAR}{}
 \item{OSF_INSPECTION_OTHER_IND}{}
 \item{OSF_INSPECTION_OTHER_YEAR}{}
 \item{OSF_INSPECTION_OTHER_DETAILS}{}
 \item{OSF_BEFORE_DAMAGE}{}
 \item{OSF_HYDROTEST_CONDUCTED_IND}{}
 \item{OSF_HYDROTEST_CONDUCTED_YEAR}{}
 \item{OSF_HYDROTEST_PRESSURE}{}
 \item{OSF_DIRECT_INSPECTION_TYPE}{}
 \item{OSF_DIRECT_YES_DIG_YEAR}{}
 \item{OSF_DIRECT_YES_NO_DIG_YEAR}{}
 \item{OSF_NON_DESTRUCTIVE_IND}{}
 \item{OSF_RADIOGRAPHY_IND}{}
 \item{OSF_RADIOGRAPHY_YEAR}{}
 \item{OSF_GUIDED_WAVE_IND}{}
 \item{OSF_GUIDED_WAVE_YEAR}{}
 \item{OSF_HANDHELD_ULTRA_IND}{}
 \item{OSF_HANDHELD_ULTRA_YEAR}{}
 \item{OSF_WET_MAGNETIC_IND}{}
 \item{OSF_WET_MAGNETIC_YEAR}{}
 \item{OSF_DRY_MAGNETIC_IND}{}
 \item{OSF_DRY_MAGNETIC_YEAR}{}
 \item{OSF_NON_DEST_OTHER_IND}{}
 \item{OSF_NON_DEST_OTHER_YEAR}{}
 \item{OSF_NON_DEST_DETAILS}{}
 \item{INTENTIONAL_SUBTYPE}{}
 \item{INTENTIONAL_DETAILS}{}
 \item{OSF_OTHER_DETAILS}{}
 \item{FIELD_EXAM_IND}{}
 \item{METALLURGICAL_IND}{}
 \item{OTHER_ANALYSIS_IND}{}
 \item{OTHER_ANALYSIS_DETAILS}{}
 \item{STILL_UNDER_INVEST_IND}{}
 \item{FAILURE_TYPE}{}
 \item{FATIGUE_VIBR_RELATED_1}{}
 \item{FAILURE_SUBTYPE_1}{}
 \item{FATIGUE_VIBR_RELATED_OTHER_1}{}
 \item{MECHANICAL_STRESS_1}{}
 \item{OTHER_FACTOR_1}{}
 \item{OTHER_FACTOR_DETAILS_1}{}
 \item{FATIGUE_VIBR_RELATED_2}{}
 \item{FAILURE_SUBTYPE_2}{}
 \item{FATIGUE_VIBR_RELATED_OTHER_2}{}
 \item{MECHANICAL_STRESS_2}{}
 \item{OTHER_FACTOR_2}{}
 \item{OTHER_FACTOR_DETAILS_2}{}
 \item{STRESS_SUBTYPE}{}
 \item{STRESS_DETAILS}{}
 \item{ADDITIONAL_DENT_IND}{}
 \item{ADDITIONAL_GOUGE_IND}{}
 \item{ADDITIONAL_PIPE_BEND_IND}{}
 \item{ADDITIONAL_ARC_BURN_IND}{}
 \item{ADDITIONAL_CRACK_IND}{}
 \item{ADDITIONAL_LACK_FUSION_IND}{}
 \item{ADDITIONAL_LAMINATION_IND}{}
 \item{ADDITIONAL_BUCKLE_IND}{}
 \item{ADDITIONAL_WRINKLE_IND}{}
 \item{PWF_ADDL_MISALIGNMENT_IND}{}
 \item{ADDITIONAL_BURNT_STEEL_IND}{}
 \item{PWF_ADDITIONAL_OTHER_IND}{}
 \item{PWF_ADDITIONAL_OTHER_DETAILS}{}
 \item{PWF_INSP_TOOL_COLLECTED_IND}{}
 \item{PWF_MAGNETIC_FLUX_LEAKAGE_IND}{}
 \item{PWF_MAG_FLUX_LEAKAGE_YEAR}{}
 \item{PWF_ULTRASONIC_IND}{}
 \item{PWF_ULTRASONIC_YEAR}{}
 \item{PWF_GEOMETRY_IND}{}
 \item{PWF_GEOMETRY_YEAR}{}
 \item{PWF_CALIPER_IND}{}
 \item{PWF_CALIPER_YEAR}{}
 \item{PWF_CRACK_IND}{}
 \item{PWF_CRACK_YEAR}{}
 \item{PWF_HARDSPOT_IND}{}
 \item{PWF_HARDSPOT_YEAR}{}
 \item{PWF_COMBINATION_TOOL_IND}{}
 \item{PWF_COMBINATION_TOOL_YEAR}{}
 \item{PWF_TRANSVERSE_FIELD_IND}{}
 \item{PWF_TRANSVERSE_FIELD_YEAR}{}
 \item{PWF_INSPECTION_OTHER_IND}{}
 \item{PWF_INSPECTION_OTHER_YEAR}{}
 \item{PWF_INSPECTION_OTHER_DETAILS}{}
 \item{PWF_HYDROTEST_CONDUCTED_IND}{}
 \item{PWF_HYDROTEST_CONDUCTED_YEAR}{}
 \item{PWF_HYDROTEST_PRESSURE}{}
 \item{PWF_DIRECT_INSPECTION_TYPE}{}
 \item{PWF_DIRECT_YES_DIG_YEAR}{}
 \item{PWF_DIRECT_YES_NO_DIG_YEAR}{}
 \item{PWF_NON_DEST_IND}{}
 \item{PWF_RADIOGRAPHY_IND}{}
 \item{PWF_RADIOGRAPHY_YEAR}{}
 \item{PWF_GUIDED_WAVE_IND}{}
 \item{PWF_GUIDED_WAVE_YEAR}{}
 \item{PWF_HANDHELD_ULTRA_IND}{}
 \item{PWF_HANDHELD_ULTRA_YEAR}{}
 \item{PWF_WET_MAGNETIC_IND}{}
 \item{PWF_WET_MAGNETIC_YEAR}{}
 \item{PWF_DRY_MAGNETIC_IND}{}
 \item{PWF_DRY_MAGNETIC_YEAR}{}
 \item{PWF_NON_DEST_OTHER_IND}{}
 \item{PWF_NON_DEST_OTHER_YEAR}{}
 \item{PWF_NON_DEST_OTHER_DETAILS}{}
 \item{EQ_FAILURE_TYPE}{}
 \item{CONTROL_VALVE_IND}{}
 \item{INSTRUMENTATION_IND}{}
 \item{SCADA_IND}{}
 \item{COMMUNICATIONS_IND}{}
 \item{BLOCK_VALVE_IND}{}
 \item{CHECK_VALVE_IND}{}
 \item{RELIEF_VALVE_IND}{}
 \item{POWER_FAILURE_IND}{}
 \item{STOPPLE_CONTROL_FITTING_IND}{}
 \item{ESD_SYSTEM_FAILURE_IND}{}
 \item{OTHER_CONTROL_RELIEF_IND}{}
 \item{OTHER_CONTROL_RELIEF_DETAILS}{}
 \item{OTHER_PUMP_IND}{}
 \item{OTHER_PUMP_DETAILS}{}
 \item{OTHER_STRIPPED_IND}{}
 \item{OTHER_STRIPPED_DETAILS}{}
 \item{OTHER_NON_THREADED_IND}{}
 \item{OTHER_NON_THREADED_DETAILS}{}
 \item{FAILURE_DETAILS}{}
 \item{ADDITIONAL_VIBRATION_IND}{}
 \item{ADDITIONAL_OVERPRESSURE_IND}{}
 \item{ADDITIONAL_SUPPORT_IND}{}
 \item{ADDITIONAL_DEFECT_IND}{}
 \item{ADDITIONAL_ELECTRICITY_IND}{}
 \item{ADDITIONAL_INSTALLATION_IND}{}
 \item{ADDITIONAL_MISMATCH_IND}{}
 \item{ADDITIONAL_DISSIMILAR_IND}{}
 \item{ADDITIONAL_BREAKDOWN_IND}{}
 \item{ADDITIONAL_VALVE_IND}{}
 \item{ADDITIONAL_ALARM_IND}{}
 \item{IEF_ADDL_MISALIGNMENT_IND}{}
 \item{ADDITIONAL_THERMAL_IND}{}
 \item{EQ_ADDITIONAL_OTHER_IND}{}
 \item{EQ_ADDITIONAL_OTHER_DETAILS}{}
 \item{OPERATION_TYPE}{}
 \item{OVERFLOW_OTHER_IND}{}
 \item{OVERFLOW_OTHER_DETAILS}{}
 \item{OPERATION_DETAILS}{}
 \item{RELATED_INADEQUATE_PROC_IND}{}
 \item{RELATED_NO_PROC_IND}{}
 \item{RELATED_FAILURE_FOLLOW_IND}{}
 \item{RELATED_OTHER_IND}{}
 \item{OPERATION_RELATED_DETAILS}{}
 \item{CATEGORY_TYPE}{}
 \item{OPERATOR_QUALIFICATION_IND}{}
 \item{QUALIFIED_INDIVIDUALS}{}
 \item{OTHER_TYPE}{}
 \item{MISC_DETAILS}{}
 \item{UNKNOWN_SUBTYPE}{}
 \item{PREPARER_NAME}{}
 \item{PREPARER_TITLE}{}
 \item{PREPARER_EMAIL}{}
 \item{PREPARER_TELEPHONE}{}
 \item{PREPARER_FAX}{}
 \item{PREPARED_DATE}{}
 \item{AUTHORIZER_NAME}{}
 \item{AUTHORIZER_TITLE}{}
 \item{AUTHORIZER_TELEPHONE}{}
 \item{AUTHORIZER_EMAIL}{}
 \item{narrative}{Written description of the incident.}
 \item{NET_LOSS_BBLS}{}
 \item{EST_COST_OPER_PAID_CURRENT}{}
 \item{EST_COST_GAS_RELEASED_CURRENT}{}
 \item{EST_COST_PROP_DAMAGE_CURRENT}{}
 \item{EST_COST_EMERGENCY_CURRENT}{}
 \item{EST_COST_ENVIRONMENTAL_CURRENT}{}
 \item{EST_COST_OTHER_CURRENT}{}
 \item{TOTAL_COST_IN84}{}
 \item{TOTAL_COST_CURRENT}{}
 \item{MAP_CAUSE}{}
 \item{MAP_SUBCAUSE}{}
 \item{SPILL_TYPE_CATEGORY}{}
}}
\source{
United States Department of Transportation (DOT) Pipeline and
  Hazardous Materials Safety Administration (PHMSA).
  \url{https://www.phmsa.dot.gov/data-and-statistics/pipeline/distribution-transmission-gathering-lng-and-liquid-accident-and-incident-data}
   (right side bar).
}
\usage{
incidents_2010
}
\description{
Distribution, Transmission & Gathering, LNG, and Liquid Accident and Incident Data.
}
\details{
Title 49 of the Code of Federal Regulations (49 CFR Parts 191, 195) requires
pipeline operators to submit incident reports within 30 days of a pipeline
incident or accident. The CFR defines accidents and incidents, as well as
criteria for submitting reports to the Office of Pipeline Safety. The
following kinds of information are collected:
* Key report information
* In-depth location information
* Facility information
* Operating information
* Drug and Alcohol information
* Cause of the accident/incident
Specific information includes the time and location of the incident(s),
number of any injuries and/ or fatalities, commodity spilled/gas released,
causes of failure and evacuation procedures. The reports are used for
identifying long- and short-term trends at the national, state and
operator-specific levels. The frequency, causes, and consequences of the
incidents provide insight into the safety metrics currently used by PHMSA,
state partners, and other pipeline safety stakeholders, including the
pipeline industry and general public. PHMSA also uses the data for inspection
planning and risk assessment.
}
\examples{
\dontrun{
 incidents_2010
}
}
\keyword{datasets}
