% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\docType{data}
\name{labels}
\alias{labels}
\title{Labels for the 23 topics}
\format{
A data frame with one row for each of the 23 topics.
\describe{
  \item{topic}{Topic 1 through 23.}
  \item{label}{Label for the topic.}
  \item{cause_topic}{Whether this label constitutes a description of the spill cause.}
}
}
\usage{
labels
}
\description{
Labels for the 23 topics
}
\keyword{datasets}
