% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\docType{data}
\name{pipelines}
\alias{pipelines}
\title{Annual data on pipelines in the US from 2004 to 2018}
\format{A data frame with columns:
\describe{
 \item{year}{Report year.}
 \item{commodity}{The type of commodity being transported (Crude, HVL, etc.).}
 \item{ID}{Unique operator ID.}
 \item{name}{Name of the pipelines operator at the time of the report year.}
 \item{hca_total}{Miles of pipelines in High Consequences Areas.}
 \item{hca_offshore}{Miles of pipelines offshore in High Consequences Areas.}
 \item{hca_onshore}{Miles of pipelines onshore in High Consequences Areas.}
 \item{total_miles}{Miles of pipelines onshore.}
 \item{total_onshore}{Miles of pipelines onshore.}
 \item{total_offshore}{Miles of pipelines offshore.}
 \item{incidents}{Total number of incidents (both significant and non-significant).}
 \item{significant_incidents}{Count of significant incidents. Obtained from the incident datasets.}
 \item{serious_incidents}{Count of serious incidents. Obtained from the incident datasets.}
 \item{incidents_volume}{Total number of gallons spilled across all incidents.}
 \item{significant_incidents_volume}{Total number of gallons spilled across all significant incidents.}
 \item{incidents_cost}{Combined cost of all incidents.}
 \item{significant_incidents_cost}{Combined cost of all significant incidents.}
}}
\source{
United States Department of Transportation (DOT) Pipeline and
  Hazardous Materials Safety Administration (PHMSA).
  \url{https://www.phmsa.dot.gov/data-and-statistics/pipeline/gas-distribution-gas-gathering-gas-transmission-hazardous-liquids}
   (right side bar).
}
\usage{
pipelines
}
\description{
Annual Report Data from Gas Distribution, Gas Gathering, Gas Transmission,
Hazardous Liquids, Liquefied Natural Gas (LNG), and Underground Natural Gas
Storage (UNGS) Facility Operators. Operators are required to submit annual
reports to PHMSA. Reports include information such as total pipeline mileage,
facilities, commodities transported, mileage by material, and installation
dates. Organizations aggregated according to the M&A dataset in this package.
}
\examples{
\dontrun{
 pipelines
}
}
\keyword{datasets}
