% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines_2004.R
\docType{data}
\name{pipelines_2004}
\alias{pipelines_2004}
\title{Annual data on pipelines in the US from 2004 to 2010}
\format{A data frame with columns:
\describe{
 \item{ORIGINAL}{}
 \item{SUPPLEMENT}{}
 \item{year}{Report year.}
 \item{RPTID}{}
 \item{HL_SUP_ID}{}
 \item{commodity}{The type of commodity being transported (Crude, HVL, etc.).}
 \item{ID}{Unique operator ID.}
 \item{name}{Name of the pipelines operator at the time of the report year.}
 \item{PARENT}{}
 \item{OFSTREET}{}
 \item{OFCITY}{}
 \item{OFCOUNTY}{}
 \item{OFSTATE}{}
 \item{OFZIP}{}
 \item{HQSTREET}{}
 \item{HQCITY}{}
 \item{HQCOUNTY}{}
 \item{HQSTATE}{}
 \item{HQZIP}{}
 \item{CPBMT}{}
 \item{CPBOFFM}{}
 \item{CPBONM}{}
 \item{CPCMT}{}
 \item{CPCOFFM}{}
 \item{CPCONM}{}
 \item{CUBOFFM}{}
 \item{CUBONM}{}
 \item{CUCMT}{}
 \item{CUCOFFM}{}
 \item{CUCONM}{}
 \item{CUPMT}{}
 \item{hca_total}{Miles of pipelines in High Consequences Areas.}
 \item{hca_offshore}{Miles of pipelines offshore in High Consequences Areas.}
 \item{hca_onshore}{Miles of pipelines onshore in High Consequences Areas.}
 \item{STOFFM_1}{}
 \item{STOFFM_2}{}
 \item{STOFFM_3}{}
 \item{STOFFM_4}{}
 \item{STOFFM_5}{}
 \item{STOFFM_6}{}
 \item{STOFFM_7}{}
 \item{STOFFM_8}{}
 \item{STOFFM_9}{}
 \item{STOFFM_10}{}
 \item{STOFFM_11}{}
 \item{STOFFM_12}{}
 \item{STOFFM_13}{}
 \item{STOFFM_14}{}
 \item{STOFFM_15}{}
 \item{STOFFM_16}{}
 \item{STOFFM_17}{}
 \item{STOFFM_18}{}
 \item{STONM_1}{}
 \item{STONM_2}{}
 \item{STONM_3}{}
 \item{STONM_4}{}
 \item{STONM_5}{}
 \item{STONM_6}{}
 \item{STONM_7}{}
 \item{STONM_8}{}
 \item{STONM_9}{}
 \item{STONM_10}{}
 \item{STONM_11}{}
 \item{STONM_12}{}
 \item{STONM_13}{}
 \item{STONM_14}{}
 \item{STONM_15}{}
 \item{STONM_16}{}
 \item{STONM_17}{}
 \item{STONM_18}{}
 \item{DINSTM_1}{}
 \item{DINSTM_2}{}
 \item{DINSTM_3}{}
 \item{DINSTM_4}{}
 \item{DINSTM_5}{}
 \item{DINSTM_6}{}
 \item{DINSTM_7}{}
 \item{DINSTM_8}{}
 \item{DINSTM_9}{}
 \item{DINSTM_10}{}
 \item{miles_total}{Total miles of pipelines.}
 \item{ERWHM_1}{}
 \item{ERWHM_2}{}
 \item{ERWHM_3}{}
 \item{ERWHM_4}{}
 \item{ERWHM_5}{}
 \item{ERWHM_6}{}
 \item{ERWHM_7}{}
 \item{ERWHM_8}{}
 \item{ERWHMT}{}
 \item{ERWLM_1}{}
 \item{ERWLM_2}{}
 \item{ERWLM_3}{}
 \item{ERWLM_4}{}
 \item{ERWLM_5}{}
 \item{ERWLM_6}{}
 \item{ERWLM_7}{}
 \item{ERWLM_8}{}
 \item{ERWLMT}{}
 \item{ERWTM_1}{}
 \item{ERWTM_2}{}
 \item{ERWTM_3}{}
 \item{ERWTM_4}{}
 \item{ERWTM_5}{}
 \item{ERWTM_6}{}
 \item{ERWTM_7}{}
 \item{ERWTM_8}{}
 \item{ERWTMT}{}
 \item{YSGOFFM}{}
 \item{YSGONM}{}
 \item{YSLOFFM}{}
 \item{YSLONM}{}
 \item{BTNPMS}{}
 \item{REGGATHM}{}
 \item{BT1COM}{}
 \item{BT1NUM1}{}
 \item{BT1NUM2}{}
 \item{BT1NUM3}{}
 \item{BT1NUM4}{}
 \item{BT1NUMT}{}
 \item{BT2COM}{}
 \item{BT2NUM1}{}
 \item{BT2NUM2}{}
 \item{BT2NUM3}{}
 \item{BT2NUM4}{}
 \item{BT2NUMT}{}
 \item{volume_crude_total}{Total volume of crude oil transported on- and offshore (in barrel-miles).}
 \item{volume_hvl_total}{Total volume of highly volatile liquid transported on- and offshore (in barrel-miles).}
 \item{VTM_3}{}
 \item{volume_rpp_total}{Total volume of refined petroleum products transported on- and offshore (in barrel-miles).}
 \item{volume_other_total}{Total volume of co2 and other products transported on- and offshore (in barrel-miles).}
 \item{VTM_6}{}
 \item{II1_1}{}
 \item{II1_2}{}
 \item{II1_3}{}
 \item{II1_4}{}
 \item{II1_T}{}
 \item{II2_1}{}
 \item{II2_2}{}
 \item{II2_3}{}
 \item{II2_4}{}
 \item{II2_5}{}
 \item{II3_1}{}
 \item{II3_2}{}
 \item{II3_3}{}
 \item{II3_4}{}
 \item{II4_1}{}
 \item{II4_2}{}
 \item{II4_3}{}
 \item{II4_4}{}
 \item{II4_5}{}
 \item{II5_1}{}
 \item{II5_2}{}
 \item{BA_1}{}
 \item{BA_2}{}
 \item{BA_3}{}
 \item{BA_4}{}
 \item{BA_5}{}
 \item{BA_6}{}
 \item{BA_7}{}
 \item{BA_8}{}
 \item{DOR}{}
 \item{DOE}{}
 \item{PNAME}{}
 \item{PPHONE}{}
 \item{PEMAIL}{}
 \item{PFAX}{}
 \item{miles_onshore}{Miles of pipelines onshore (calculated from part B - miles of steel by location/protection).}
 \item{miles_offshore}{Miles of pipelines offshore (calculated from part B - miles of steel by location/protection).}
 \item{offshore_share}{Share of total pipeline miles that is offshore pipelines.}
 \item{volume_crude_offshore}{Volume of crude oil transported offshore (in barrel-miles).}
 \item{volume_crude_onshore}{Volume of crude oil transported onshore (in barrel-miles).}
 \item{volume_hvl_offshore}{Volume of highly volatile liquid transported offshore (in barrel-miles).}
 \item{volume_hvl_onshore}{Volume of highly volatile liquid transported onshore (in barrel-miles).}
 \item{volume_rpp_offshore}{Volume of refined petroleum products transported offshore (in barrel-miles).}
 \item{volume_rpp_onshore}{Volume of refined petroleum products transported onshore (in barrel-miles).}
 \item{volume_other_offshore}{Volume of co2 and other products transported offshore (in barrel-miles).}
 \item{volume_other_onshore}{Volume of co2 and other products transported onshore (in barrel-miles).}
 \item{estimate_volume_crude_offshore}{Estimated volume of crude oil transported offshore (in barrel-miles).}
 \item{estimate_volume_crude_onshore}{Estimated volume of crude oil transported onshore (in barrel-miles).}
 \item{estimate_volume_hvl_offshore}{Estimated volume of highly volatile liquid transported offshore (in barrel-miles).}
 \item{estimate_volume_hvl_onshore}{Estimated volume of highly volatile liquid transported onshore (in barrel-miles).}
 \item{estimate_volume_rpp_offshore}{Estimated volume of refined petroleum products transported offshore (in barrel-miles).}
 \item{estimate_volume_rpp_onshore}{Estimated volume of refined petroleum products transported onshore (in barrel-miles).}
 \item{estimate_volume_other_offshore}{Estimated volume of co2 and other products transported offshore (in barrel-miles).}
 \item{estimate_volume_other_onshore}{Estimated volume of co2 and other products transported onshore (in barrel-miles).}
 \item{volume_all_total}{Combined volume of all substances transported (in barrel-miles).}
 \item{volume_all_offshore}{Combined volume of all substances transported offshore (in barrel-miles).}
 \item{volume_all_onshore}{combined volume of all substances transported onshore (in barrel-miles).}
 \item{estimate_volume_all_offshore}{Estimated volume of all substances transported offshore (in barrel-miles).}
 \item{estimate_volume_all_onshore}{Estimated volume of all substances transported onshore (in barrel-miles).}
 \item{estimate_volume_all_total}{Estimated of volume of all substances transporte (in barrel-miles). Placeholder - contains the actual value as found in volume_all_total.}
}}
\source{
United States Department of Transportation (DOT) Pipeline and
  Hazardous Materials Safety Administration (PHMSA).
  \url{https://www.phmsa.dot.gov/data-and-statistics/pipeline/gas-distribution-gas-gathering-gas-transmission-hazardous-liquids}
   (right side bar).
}
\usage{
pipelines_2004
}
\description{
Annual Report Data from Gas Distribution, Gas Gathering, Gas Transmission,
Hazardous Liquids, Liquefied Natural Gas (LNG), and Underground Natural Gas
Storage (UNGS) Facility Operators. Operators are required to submit annual
reports to PHMSA. Reports include information such as total pipeline mileage,
facilities, commodities transported, mileage by material, and installation
dates.
}
\details{
The volume transported on- and offshore is created from reporting of total
transported. Where the share of offshore pipeline miles is 0%, the original
value of total volume transported is provided for onshore, and 0 is reported
for offshore. If an organization also operates offshore pipelines, NA is
recorded. Estimate_volume provides an estimate that is calculated as share
offshore multiplied by total volume for offshore; and 1 - share offshore
multiplied by total volume is used to calculate the estimate for the onshore
volume.
}
\examples{
\dontrun{
 pipelines_2004
}
}
\keyword{datasets}
