% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pipelines_2010}
\alias{pipelines_2010}
\title{Annual data on pipelines in the US from 2010 onward}
\format{A data frame with columns:
\describe{
 \item{DATAFILE_AS_OF}{}
 \item{FILING_DATE}{}
 \item{FORM_REV}{}
 \item{ID}{Unique operator ID.}
 \item{Name}{Name of the pipelines operator at the time of the report year.}
 \item{PARTA4CITY}{}
 \item{PARTA4COUNTRY}{}
 \item{PARTA4STATE}{}
 \item{PARTA4STREET}{}
 \item{PARTA4ZIP}{}
 \item{PARTA5COMMODITY}{}
 \item{PARTA7INTER}{}
 \item{PARTA7INTRA}{}
 \item{PARTBHCAOFFSHORE}{}
 \item{PARTBHCAONSHORE}{}
 \item{PARTBHCATOTAL}{}
 \item{PARTCOFFCO2}{}
 \item{PARTCOFFCRUDE}{}
 \item{PARTCOFFETHANOL}{}
 \item{PARTCOFFHVL}{}
 \item{PARTCOFFRPP}{}
 \item{PARTCONCO2}{}
 \item{PARTCONCRUDE}{}
 \item{PARTCONETHANOL}{}
 \item{PARTCONHVL}{}
 \item{PARTCONRPP}{}
 \item{PARTDCPBTOTAL}{}
 \item{PARTDCPCTOTAL}{}
 \item{PARTDCUBTOTAL}{}
 \item{PARTDCUCTOTAL}{}
 \item{PARTDCUOTOTAL}{}
 \item{PARTDCUPTOTAL}{}
 \item{PARTDOFFCPB}{}
 \item{PARTDOFFCPC}{}
 \item{PARTDOFFCUB}{}
 \item{PARTDOFFCUC}{}
 \item{PARTDOFFCUO}{}
 \item{PARTDOFFCUP}{}
 \item{PARTDOFFTOTAL}{}
 \item{PARTDONCPB}{}
 \item{PARTDONCPC}{}
 \item{PARTDONCUB}{}
 \item{PARTDONCUC}{}
 \item{PARTDONCUO}{}
 \item{PARTDONCUP}{}
 \item{PARTDONTOTAL}{}
 \item{PARTDTOTALMILES}{}
 \item{PARTE1940HF}{}
 \item{PARTE1940LF}{}
 \item{PARTE1940TOTAL}{}
 \item{PARTE1950HF}{}
 \item{PARTE1950LF}{}
 \item{PARTE1950TOTAL}{}
 \item{PARTE1960HF}{}
 \item{PARTE1960LF}{}
 \item{PARTE1960TOTAL}{}
 \item{PARTE1970HF}{}
 \item{PARTE1970LF}{}
 \item{PARTE1970TOTAL}{}
 \item{PARTE1980HF}{}
 \item{PARTE1980LF}{}
 \item{PARTE1980TOTAL}{}
 \item{PARTE1990HF}{}
 \item{PARTE1990LF}{}
 \item{PARTE1990TOTAL}{}
 \item{PARTE2000HF}{}
 \item{PARTE2000LF}{}
 \item{PARTE2000TOTAL}{}
 \item{PARTE2010HF}{}
 \item{PARTE2010LF}{}
 \item{PARTE2010TOTAL}{}
 \item{PARTEPRE40HF}{}
 \item{PARTEPRE40LF}{}
 \item{PARTEPRE40TOTAL}{}
 \item{PARTETOTAL}{}
 \item{PARTETOTALHF}{}
 \item{PARTETOTALLF}{}
 \item{PARTEUNKNHF}{}
 \item{PARTEUNKNLF}{}
 \item{PARTEUNKNTOTAL}{}
 \item{REPORT_DATE}{}
 \item{REPORT_NUMBER}{}
 \item{REPORT_SUBMISSION_TYPE}{}
 \item{REPORT_YEAR}{}
 \item{SUPPLEMENTAL_NUMBER}{}
}}
\source{
United States Department of Transportation (DOT) Pipeline and
  Hazardous Materials Safety Administration (PHMSA).
  \url{https://www.phmsa.dot.gov/data-and-statistics/pipeline/gas-distribution-gas-gathering-gas-transmission-hazardous-liquids}
   (right side bar).
}
\usage{
pipelines_2010
}
\description{
Annual Report Data from Gas Distribution, Gas Gathering, Gas Transmission,
Hazardous Liquids, Liquefied Natural Gas (LNG), and Underground Natural Gas
Storage (UNGS) Facility Operators. Operators are required to submit annual
reports to PHMSA. Reports include information such as total pipeline mileage,
facilities, commodities transported, mileage by material, and installation
dates.
}
\examples{
\dontrun{
 pipelines_2010
}
}
\keyword{datasets}
