% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines_2010.R
\docType{data}
\name{pipelines_2010}
\alias{pipelines_2010}
\title{Annual data on pipelines in the US from 2010 onward}
\format{A data frame with columns:
\describe{
 \item{DATAFILE_AS_OF}{}
 \item{year}{Report year.}
 \item{REPORT_NUMBER}{}
 \item{SUPPLEMENTAL_NUMBER}{}
 \item{ID}{Unique operator ID.}
 \item{name}{Name of the pipelines operator at the time of the report year.}
 \item{PARTA4STREET}{}
 \item{PARTA4CITY}{}
 \item{PARTA4STATE}{}p
 \item{PARTA4ZIP}{}
 \item{PARTA4COUNTRY}{}
 \item{commodity}{The type of commodity being transported (Crude, HVL, etc.).}
 \item{PARTA7INTER}{}
 \item{PARTA7INTRA}{}
 \item{hca_offshore}{Miles of pipelines offshore in High Consequences Areas.}
 \item{hca_onshore}{Miles of pipelines onshore in High Consequences Areas.}
 \item{hca_total}{Miles of pipelines in High Consequences Areas.}
 \item{PARTCOFFCO2}{}
 \item{PARTCOFFCRUDE}{}
 \item{PARTCOFFETHANOL}{}
 \item{PARTCOFFHVL}{}
 \item{PARTCOFFRPP}{}
 \item{PARTCONCO2}{}
 \item{PARTCONCRUDE}{}
 \item{PARTCONETHANOL}{}
 \item{PARTCONHVL}{}
 \item{PARTCONRPP}{}
 \item{PARTDONCPB}{}
 \item{PARTDONCPC}{}
 \item{PARTDONCUB}{}
 \item{PARTDONCUC}{}
 \item{PARTDONCUP}{}
 \item{PARTDONCUO}{}
 \item{total_onshore}{Miles of pipelines onshore.}
 \item{PARTDOFFCPB}{}
 \item{PARTDOFFCPC}{}
 \item{PARTDOFFCUB}{}
 \item{PARTDOFFCUC}{}
 \item{PARTDOFFCUP}{}
 \item{PARTDOFFCUO}{}
 \item{total_offshore}{Miles of pipelines offshore.}
 \item{PARTDCPBTOTAL}{}
 \item{PARTDCPCTOTAL}{}
 \item{PARTDCUBTOTAL}{}
 \item{PARTDCUCTOTAL}{}
 \item{PARTDCUPTOTAL}{}
 \item{PARTDCUOTOTAL}{}
 \item{total_miles}{Total miles of pipelines.}
 \item{PARTEUNKNHF}{}
 \item{PARTEUNKNLF}{}
 \item{PARTEUNKNTOTAL}{}
 \item{PARTEPRE40HF}{}
 \item{PARTEPRE40LF}{}
 \item{PARTEPRE40TOTAL}{}
 \item{PARTE1940HF}{}
 \item{PARTE1940LF}{}
 \item{PARTE1940TOTAL}{}
 \item{PARTE1950HF}{}
 \item{PARTE1950LF}{}
 \item{PARTE1950TOTAL}{}
 \item{PARTE1960HF}{}
 \item{PARTE1960LF}{}
 \item{PARTE1960TOTAL}{}
 \item{PARTE1970HF}{}
 \item{PARTE1970LF}{}
 \item{PARTE1970TOTAL}{}
 \item{PARTE1980HF}{}
 \item{PARTE1980LF}{}
 \item{PARTE1980TOTAL}{}
 \item{PARTE1990HF}{}
 \item{PARTE1990LF}{}
 \item{PARTE1990TOTAL}{}
 \item{PARTE2000HF}{}
 \item{PARTE2000LF}{}
 \item{PARTE2000TOTAL}{}
 \item{PARTE2010HF}{}
 \item{PARTE2010LF}{}
 \item{PARTE2010TOTAL}{}
 \item{PARTETOTAL}{}
 \item{PARTETOTALHF}{}
 \item{PARTETOTALLF}{}
 \item{REPORT_SUBMISSION_TYPE}{}
 \item{REPORT_DATE}{}
 \item{FILING_DATE}{}
 \item{FORM_REV}{}
}}
\source{
United States Department of Transportation (DOT) Pipeline and
  Hazardous Materials Safety Administration (PHMSA).
  \url{https://www.phmsa.dot.gov/data-and-statistics/pipeline/gas-distribution-gas-gathering-gas-transmission-hazardous-liquids}
   (right side bar).
}
\usage{
pipelines_2010
}
\description{
Annual Report Data from Gas Distribution, Gas Gathering, Gas Transmission,
Hazardous Liquids, Liquefied Natural Gas (LNG), and Underground Natural Gas
Storage (UNGS) Facility Operators. Operators are required to submit annual
reports to PHMSA. Reports include information such as total pipeline mileage,
facilities, commodities transported, mileage by material, and installation
dates.
}
\examples{
\dontrun{
 pipelines_2010
}
}
\keyword{datasets}
